% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_connection.R
\name{show_connection}
\alias{show_connection}
\title{A Cat Function}
\usage{
show_connection(o1, o2, tokeep = NA)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
