% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LCIS}
\alias{LCIS}
\title{Longuest Common Induced qSequence (LCIS)}
\usage{
LCIS(order_1, order_2)
}
\arguments{
\item{order_1}{the first list. Must be 2 columns called V1 and V2. V1 being the name
of the item, V2 its position}

\item{order_2}{the second list. Must be 2 columns called V1 and V2. V1 being the name
of the item, V2 its position}
}
\value{
The function returns a list with several informations:
		- $order_1: the first list
		- $order_2: the second list
		- $LLCIS: the number of items in the longuest common sequence
		- $LCIS: the item names in the longuest common sequence
		- $QSI: an estimate of the proportion of items kept in the LCS
		- $LCIS_pos_order_1: the position of kept items in the first list
		- $LCIS_pos_order_2: the position of kept items in the second list
}
\description{
Find the Longuest Common Induced Sequence (LCIS) between 2 lists.
}
\details{
The LCS function takes 2 list of items and return the biggest subset of items
showing the same order. If several items are at the same position, only one will
be kept per position.
}
\examples{
# create 2 lists of items, each item having a position
o1 <- data.frame( 
   	V1=c("g","h","c","f","e","d","m","q","r","a","b","n","o","p"), 
   	V2=c(1.1,1.1,3.4,3.4,3.4,3.4,3.5,3.5,6.6,6.6,7.2,7.2,8,8)
   )
o2 <- data.frame( 
	V1=c("k","a","b","l","c","e","d","f","i","j","h","g"), 
   	V2=c(0.1,1.2,1.2,4.2,4.2,5,5,5,5.3,5.3,6.7,6.7)
   )
# run LCS
LCIS(o1, o2)
}
\author{
Lisa de Matteo
}
