% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_connection.R
\name{show_connection}
\alias{show_connection}
\title{A function to draw the relationship between 2 (bucket) orders}
\usage{
show_connection(order_1, order_2, tokeep = NA, showName = TRUE,
  highlight = NA)
}
\arguments{
\item{order_1}{the first order, provided as a dataframe made of two columns named V1 and V2; V1 contains marker names while V2 contains their position}

\item{order_2}{the second order, provided as a dataframe made of two columns named V1 and V2; V1 contains marker names while V2 contains their position}

\item{tokeep}{a vector containg the names of the markers to be kept for this graphic, other markers are disgarded. By default, all markers are kept.}

\item{showName}{a boolean indicating whether or not marker names should be displayed (for large datasets displaying marker names often result in overly cluttered charts)}

\item{highlight}{a vector containg the names of the markers to be highlight for this graphic. The lines linking those markers will be drawn in dark grey, while others (e.g. those not part of the LCS) will be drawn in red instead of being masked as with the tokeep option.}
}
\description{
This function takes 2 marker orders order_1 and order_2 and build a diagram to ease their comparison. 
The two orders are represented by two parallel vertical lines, and marker are positoned along those lines based on their positions within order_1 and order_2. Every time a marker is present in both order_1 and order_2 (same name) a new line is drawn between: the position of this marker on the vertical line representing order_1 and its position on order_2 (by default a different color is used for each marker).
This function requires the tidyverse and ggrepel packages.
}
\examples{
# create 2 orders of markers. Each order is represented by a dataframe made of two columns, the first (V1) contains marker names (e.g, a, e, f) the second (V2) contains marker position in this order (e.g 1.1, 5, 6.7, 8) 
order_1 <- data.frame( 
   V1=c("g","h","c","f","e","d","m","q","r","a","b","n","o","p"), 
   V2=c(1.1,1.1,3.4,3.4,3.4,3.4,3.5,3.5,6.6,6.6,7.2,7.2,8,8)
   )
order_2 <- data.frame( 
   V1=c("k","a","b","l","c","e","d","f","i","j","h","g"), 
   V2=c(0.1,1.2,1.2,4.2,4.2,5,5,5,5.3,5.3,6.7,6.7)
   )

# show the connection between the markers of order_1 and order_2
library(tidyverse)
library(ggrepel)
show_connection(order_1,order_2)

# Run LCS and plot only the connection between order_1 and order_2 for their LCS markers (tokeep=res$LCS) or higlight those connections (highlight=res$LCS)
res <- LCS(order_1,order_2)
show_connection(order_1, order_2, tokeep=res$LCS)
show_connection(order_1, order_2, highlight =res$LCS)
}
\author{
Yan Holtz, Lisa de Matteo, Vincent Ranwez, Severine Berard
}
\keyword{connection}
\keyword{list}
\keyword{plot}
